+++ NEWS.md
@@ -9,7 +9,10 @@
 
 * Developer focused `.get_data_types()` generic has been added to designate types of columns. Exported for use in extension packages that deal with types not supported in recipes directly. (#993)
 
+* The `step_date()` function now defaults to using the clock package to format day-of-week and month labels. (#1048)
+
 * `step_range()` has gained a argument `clipping` that when set to `FALSE` no longer clips the data to be between `min` and `max`.
+
 
 # recipes 1.0.2